requirements.txt
arrow==0.15.7
asgiref==3.2.10
Django==3.0.8
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-filter==2.3.0
django-js-asset==1.2.2
django-userforeignkey==0.3.0
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
South==1.0.2
sqlparse==0.3.1
